+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 
 - Make pyramid_rawes compatible with rawes 0.5.x
 - Introduce unit test support for py3k (requires rawes 0.5.1)
+- Ability to specify json_encoder as dotted python name.
 
 0.3.0 (29-04-2013)
 ------------------